requirements.txt
boto3
openai
python-dotenv
fastapi
uvicorn
python-multipart
